+++ messaging-core/src/main/java/com/nabla/sdk/core/domain/entity/DeletedProvider.kt
@@ -1,0 +1,3 @@
+package com.nabla.sdk.core.domain.entity
+
+public object DeletedProvider : MaybeProvider