+++ messaging-core/src/main/java/com/nabla/sdk/messaging/core/domain/boundary/ConversationContentRepository.kt
@@ -1,13 +1,13 @@
 package com.nabla.sdk.messaging.core.domain.boundary
 
-import com.nabla.sdk.messaging.core.data.message.PaginatedConversationMessages
+import com.nabla.sdk.messaging.core.data.message.PaginatedConversationItems
 import com.nabla.sdk.messaging.core.domain.entity.ConversationId
 import com.nabla.sdk.messaging.core.domain.entity.Message
 import com.nabla.sdk.messaging.core.domain.entity.MessageId
 import kotlinx.coroutines.flow.Flow
 
-internal interface MessageRepository {
-    fun watchConversationMessages(conversationId: ConversationId): Flow<PaginatedConversationMessages>
+internal interface ConversationContentRepository {
+    fun watchConversationItems(conversationId: ConversationId): Flow<PaginatedConversationItems>
     suspend fun loadMoreMessages(conversationId: ConversationId)
     suspend fun sendMessage(message: Message)
     suspend fun retrySendingMessage(conversationId: ConversationId, localMessageId: MessageId.Local)